feature: lift the product level and refactor the all-product component in landing page
