+++ src/components/pages/landing/Landing.tsx
@@ -1,18 +1,25 @@
-import ProductList from "./product-list/ProductList";
+import { AllProducts } from "../product/AllProducts";
 import TopSellers from "./top-sellers/TopSellers";
 import Testimonials from "./testimonials/testimonials";
+import type { Product } from "../../common/product-list/sample-data/sample-data";
 
-function Landing() {
-    return(
+function Landing({
+  products,
+  updateProducts,
+}: {
+  products: Product[];
+  updateProducts: React.Dispatch<React.SetStateAction<Product[]>>;
+}) {
+  return (
     <main>
-        {/* render the featured product list */}
-        <ProductList />
-        {/* render list of sellers with descending reputation */}
-        <TopSellers />
-        {/* render the testimonial slider */}
-        <Testimonials />
-      </main>
-    );
+      {/* render the featured product list */}
+      <AllProducts products={products} updateProducts={updateProducts} />
+      {/* render list of sellers with descending reputation */}
+      <TopSellers />
+      {/* render the testimonial slider */}
+      <Testimonials />
+    </main>
+  );
 }
 
 export default Landing;
